+++ src/ladb_opencutlist/ruby/lib/kuix/entity/3d/points.rb
@@ -28,9 +28,7 @@ module Ladb::OpenCutList::Kuix
     # -- LAYOUT --
 
     def do_layout(transformation)
-      @_points = @points.map { |point|
-        point.transform(transformation * @transformation)
-      }
+      @_points = @points.map { |point| point.transform(transformation * @transformation) }
       super
     end