+++ src/ladb_opencutlist/ruby/lib/kuix/entity/3d/segments.rb
@@ -27,9 +27,7 @@ module Ladb::OpenCutList::Kuix
     # -- LAYOUT --
 
     def do_layout(transformation)
-      @_points = @segments.map { |point|
-        point.transform(transformation * @transformation)
-      }
+      @_points = @segments.map { |point| point.transform(transformation * @transformation) }
       super
     end