+++ src/ladb_opencutlist/ruby/lib/kuix/geom/anchor.rb
@@ -2,15 +2,15 @@ module Ladb::OpenCutList::Kuix
 
   class Anchor
 
-    TOP_LEFT = 0
-    TOP_CENTER = 1
-    TOP_RIGHT = 2
-    CENTER_RIGHT = 3
-    BOTTOM_RIGHT = 4
-    BOTTOM_CENTER = 5
-    BOTTOM_LEFT = 6
-    CENTER_LEFT = 7
-    CENTER = 8
+    TOP = 1
+    LEFT = 2
+    BOTTOM = 3
+    RIGHT = 4
+    TOP_LEFT = 5
+    TOP_RIGHT = 6
+    BOTTOM_LEFT = 7
+    BOTTOM_RIGHT = 8
+    CENTER = 9
 
     attr_reader :position
 
@@ -29,27 +29,27 @@ module Ladb::OpenCutList::Kuix
     # -- Properties --
 
     def top?
-      @position == TOP_LEFT || @position == TOP_CENTER || @position == TOP_RIGHT
+      @position == TOP_LEFT || @position == TOP || @position == TOP_RIGHT
     end
 
     def right?
-      @position == TOP_RIGHT || @position == CENTER_RIGHT || @position == BOTTOM_RIGHT
+      @position == TOP_RIGHT || @position == RIGHT || @position == BOTTOM_RIGHT
     end
 
     def bottom?
-      @position == BOTTOM_LEFT || @position == BOTTOM_CENTER || @position == BOTTOM_RIGHT
+      @position == BOTTOM_LEFT || @position == BOTTOM || @position == BOTTOM_RIGHT
     end
 
     def left?
-      @position == TOP_RIGHT || @position == CENTER_LEFT || @position == BOTTOM_LEFT
+      @position == TOP_RIGHT || @position == LEFT || @position == BOTTOM_LEFT
     end
 
     def horizontal_center?
-      @position == CENTER_LEFT || @position == CENTER || @position == CENTER_RIGHT
+      @position == LEFT || @position == CENTER || @position == RIGHT
     end
 
     def vertical_center?
-      @position == TOP_CENTER || @position == CENTER || @position == BOTTOM_CENTER
+      @position == TOP || @position == CENTER || @position == BOTTOM
     end
 
     # --