+++ src/ladb_opencutlist/ruby/worker/outliner/outliner_worker.rb
@@ -119,6 +119,11 @@ module Ladb::OpenCutList
             node_def.invalidate
             node_def = node_def.parent
           end
+
+        else
+
+          puts 'active_node_def not found'
+
         end
         @outliner_def.active_node_def = active_node_def