+++ CHANGELOG.md
@@ -10,6 +10,7 @@ in 7.x, 6.x, 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
     * Added Move Axes action on Smart Axes Tool
     * Added Export Paths action on Smart Export Tool
     * Added Switch YZ axes option to Smart Export Tool on Export 3D part action
+    * Added Material type separators in materials tabs when sorted by type
     * Improved smart tools picker
     * Improved surface detection
     * Replaced Highlight Part Tool by Smart Axes Tool